+++ packages/domain-social/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - Updated dependencies []:
+  - @volontariapp/contracts@4.3.7
   - @volontariapp/messaging@2.16.0
   - @volontariapp/database@3.4.15
   - @volontariapp/outbox@0.9.52